[step3] Rank, Suit 파일 분리
