+++ src/main/kotlin/blackjack/domain/Rank.kt
@@ -16,9 +16,3 @@ enum class Rank(val value: Int, val rankName: String) {
     ACE(11, "A"), // ACE is 11, but can also be 1
 }
 
-enum class Suit(val koreanName: String) {
-    HEARTS("하트"),
-    DIAMONDS("다이아몬드"),
-    CLUBS("클로바"),
-    SPADES("스페이드"),
-}
