+++ src/main/kotlin/blackjack/domain/Suit.kt
@@ -1,0 +1,8 @@
+package blackjack.domain
+
+enum class Suit(val koreanName: String) {
+    HEARTS("하트"),
+    DIAMONDS("다이아몬드"),
+    CLUBS("클로바"),
+    SPADES("스페이드"),
+}
